+++ src/main/java/com/tangerineSpecter/oms/system/domain/vo/data/FoodLibraryInfoVo.java
@@ -31,7 +31,7 @@ public class FoodLibraryInfoVo implements Serializable {
     private String logo;
 
     @ApiModelProperty("卡路里/100g")
-    private Integer calories;
+    private BigDecimal calories;
 
     @ApiModelProperty("蛋白质/100g")
     private BigDecimal protein;
